Backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
asyncpg==0.29.0
redis==5.0.8
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.28.1
PyJWT==2.9.0
bcrypt==4.1.2
alembic==1.13.1
email-validator==2.1.1
# OCR Dependencies
google-genai==1.20.0
pdf2image==1.17.0
pillow==11.3.0
python-multipart==0.0.9

# Testing Dependencies
pytest-asyncio==0.21.1
pytz==2025.2
aiosqlite==0.21.0